implement Level 1. Echo and add enums for command
